# OCPBUGS-19056: apiserver: optimize IsLongRunningRequest