Avoid using expensive MatchString when request params are already set in requestInfo

Back in 3.x days this was necessary as buildconfigs and imagestreams was served under `/oapi`, but now they are proper CRs so its better to match them via requestInfo params

```
$ benchstat -delta-test=none old.txt new.txt               
name                                                                 old time/op  new time/op  delta
IsLongRunningRequest/path_/api-8                                     11.5ns ± 5%   9.9ns ± 8%  -14.20%
IsLongRunningRequest/path_/api/buildconfigs/foo/instantiatebinary-8   870ns ± 7%     6ns ± 9%  -99.32%
IsLongRunningRequest/path_/api/imagestreamimports-8                   391ns ± 3%    10ns ± 5%  -97.47%
```

Next this needs to be re-imported by openshift-apiserver