@benluddy 2023-10-27
Is there E2E coverage for this if you were to open a proof PR?