@vrutkovs 2023-10-27
We don't have e2e tests for this, the only thing it would affect is `apiserver_longrunning_requests` metrics iiuc