@vrutkovs 2023-11-01
Tested https://github.com/openshift/kubernetes/pull/1789 and verified metrics in e2e-aws-ovn-cgroupsv2 test:
* `apiserver_request_duration_seconds_sum{resource="buildconfigs", subresource="instantiatebinary"}` shows multiple results (longest being 18 sec), not included in `apiserver_longrunning_requests{verb!="WATCH"}`
* `apiserver_request_duration_seconds_sum{resource="imagestreamimports"}` shows multiple results (longest being 80 sec), not included in `apiserver_longrunning_requests{verb!="WATCH"}`

So it doesn't introduce any changes to long request detection