@p0lyn0mial 2024-01-03
@vrutkovs do you want to add some unit tests ? We could use `NewRequestInfo` (https://github.com/kubernetes/kubernetes/blob/bc6c7fa91201348d010b638fbadf32007c0ac546/staging/src/k8s.io/apiserver/pkg/endpoints/request/requestinfo.go#L117) to build the RequestInfo struct and pass it to the detection function.