@vrutkovs 2024-01-03
Added unittests